requirements.txt
selenium
selenium-wire
pyyaml
PyVirtualDisplay
loguru
boto3
